client/api/requirements.txt
flask
python-dotenv
spotipy
pandas
pathlib
google-cloud-datastore
google-cloud-storage
Image
werkzeug
requests
Pillow
openai